requirements.txt
wtforms
flask-wtf
werkzeug
flask-login
python-dotenv
sqlalchemy
flask-sqlalchemy
flask-migrate
flask
